mcp-servers/viz-platform/requirements.txt
# MCP SDK
mcp>=0.9.0

# Visualization
plotly>=5.18.0
dash>=2.14.0
dash-mantine-components>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.0
